+++ alphachain/test/adapter_test.js
@@ -1,0 +1,75 @@
+const { assert } = require('chai')
+const { createRequest } = require('../adapter')
+
+describe('createRequest', () => {
+  const jobID = '1'
+
+  context('successful calls', () => {
+    const requests = [
+      {
+        name: 'id not supplied',
+        testData: { data: { base: 'ETH', quote: 'XDR' } }
+      },
+      {
+        name: 'base/quote',
+        testData: { id: jobID, data: { base: 'ETH', quote: 'XDR' } }
+      },
+      {
+        name: 'from/to',
+        testData: { id: jobID, data: { from: 'ETH', to: 'XDR' } }
+      },
+      {
+        name: 'coin/market',
+        testData: { id: jobID, data: { coin: 'ETH', market: 'XDR' } }
+      }
+    ]
+
+    requests.forEach((req) => {
+      it(`${req.name}`, (done) => {
+        createRequest(req.testData, (statusCode, data) => {
+          assert.equal(statusCode, 200)
+          assert.equal(data.jobRunID, jobID)
+          assert.isNotEmpty(data.data)
+          assert.isAbove(data.result, 0)
+          assert.isAbove(data.data.result, 0)
+          done()
+        })
+      })
+    })
+  })
+
+  context('error calls', () => {
+    const requests = [
+      { name: 'empty body', testData: {} },
+      { name: 'empty data', testData: { data: {} } },
+      {
+        name: 'base not supplied',
+        testData: { id: jobID, data: { quote: 'XDR' } }
+      },
+      {
+        name: 'quote not supplied',
+        testData: { id: jobID, data: { base: 'ETH' } }
+      },
+      {
+        name: 'unknown base',
+        testData: { id: jobID, data: { base: 'not_real', quote: 'XDR' } }
+      },
+      {
+        name: 'unknown quote',
+        testData: { id: jobID, data: { base: 'ETH', quote: 'not_real' } }
+      }
+    ]
+
+    requests.forEach((req) => {
+      it(`${req.name}`, (done) => {
+        createRequest(req.testData, (statusCode, data) => {
+          assert.equal(statusCode, 500)
+          assert.equal(data.jobRunID, jobID)
+          assert.equal(data.status, 'errored')
+          assert.isNotEmpty(data.error)
+          done()
+        })
+      })
+    })
+  })
+})