+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Replaced `alphavantage-sdr` with `alphachain` adapter
+
 ## [0.1.1] - 2020-09-18
 
 ### Added